refactored to clean up codebase
Basic idea changes:
Event handlers will have the handle_events method instead of Engine.
The game map will have a reference to Engine, and entities will have a reference to the map.
Actions will be initialized with the entity doing the action
Because of the above points, Actions will have a reference to the Engine, through Entity->GameMap->Engine